requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-multipart==0.0.12
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.18.0
stripe==10.12.0

# Testing (optional, for development)
pytest>=8.0.0
